// web/src/api/generated/model/handoffOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * TCFL Portal API
 * OpenAPI spec version: 0.1.0
 */
import type { HandoffState } from './handoffState';
import type { NationalIdState } from './nationalIdState';

export interface HandoffOut {
  state: HandoffState;
  match_code: string;
  claim_expires_at: string;
  device: string | null;
  connected_at: string | null;
  national_id: NationalIdState;
  results_started: boolean;
}